Move non-storage classes out of world package
