+++ src/main/kotlin/com/github/hotm/auranet/AuraNode.kt
@@ -1,4 +1,4 @@
-package com.github.hotm.world.auranet
+package com.github.hotm.auranet
 
 import alexiil.mc.lib.net.IMsgReadCtx
 import alexiil.mc.lib.net.IMsgWriteCtx
@@ -11,6 +11,7 @@ import com.github.hotm.misc.HotMRegistries
 import com.github.hotm.mixinapi.StorageUtils
 import com.github.hotm.net.HotMNetwork
 import com.github.hotm.util.DimBlockPos
+import com.github.hotm.world.auranet.AuraNetAccess
 import com.mojang.serialization.*
 import net.minecraft.util.math.BlockPos
 import org.apache.logging.log4j.LogManager
