+++ src/main/kotlin/com/github/hotm/auranet/DependableAuraNode.kt
@@ -1,4 +1,4 @@
-package com.github.hotm.world.auranet
+package com.github.hotm.auranet
 
 import net.minecraft.util.math.BlockPos
 import java.util.stream.Stream
