+++ src/main/kotlin/com/github/hotm/auranet/PortalTXAuraNode.kt
@@ -1,4 +1,4 @@
-package com.github.hotm.world.auranet
+package com.github.hotm.auranet
 
 interface PortalTXAuraNode : AuraNode {
     fun isValid(): Boolean
