+++ src/main/kotlin/com/github/hotm/auranet/SiphonAuraNode.kt
@@ -1,4 +1,4 @@
-package com.github.hotm.world.auranet
+package com.github.hotm.auranet
 
 import com.github.hotm.util.DimBlockPos
