+++ src/main/kotlin/com/github/hotm/blocks/AbstractAuraNodeBlockWithEntity.kt
@@ -3,7 +3,7 @@ package com.github.hotm.blocks
 import alexiil.mc.lib.attributes.AttributeList
 import alexiil.mc.lib.attributes.AttributeProvider
 import com.github.hotm.mixinapi.StorageUtils
-import com.github.hotm.world.auranet.AttributeProviderAuraNode
+import com.github.hotm.auranet.AttributeProviderAuraNode
 import net.minecraft.block.BlockState
 import net.minecraft.block.BlockWithEntity
 import net.minecraft.server.world.ServerWorld
