+++ src/main/kotlin/com/github/hotm/blocks/AuraNodeBlock.kt
@@ -1,7 +1,7 @@
 package com.github.hotm.blocks
 
-import com.github.hotm.world.auranet.AuraNode
-import com.github.hotm.world.auranet.AuraNodeType
+import com.github.hotm.auranet.AuraNode
+import com.github.hotm.auranet.AuraNodeType
 import com.github.hotm.world.auranet.server.ServerAuraNetStorage
 import net.minecraft.block.BlockState
 import net.minecraft.server.world.ServerWorld
