+++ src/main/kotlin/com/github/hotm/blocks/CollectorDistributorAuraNodeBlock.kt
@@ -3,9 +3,9 @@ package com.github.hotm.blocks
 import com.github.hotm.blockentity.AbstractDependableAuraNodeBlockEntity
 import com.github.hotm.blockentity.CollectorDistributorAuraNodeBlockEntity
 import com.github.hotm.blockentity.HotMBlockEntities
-import com.github.hotm.world.auranet.AuraNode
-import com.github.hotm.world.auranet.AuraNodeType
-import com.github.hotm.world.auranet.CollectorDistributorAuraNode
+import com.github.hotm.auranet.AuraNode
+import com.github.hotm.auranet.AuraNodeType
+import com.github.hotm.auranet.CollectorDistributorAuraNode
 import com.github.hotm.world.auranet.server.ServerAuraNetStorage
 import net.minecraft.block.BlockRenderType
 import net.minecraft.block.BlockState
