+++ src/main/kotlin/com/github/hotm/blocks/PortalReceiverAuraNodeBlock.kt
@@ -3,10 +3,10 @@ package com.github.hotm.blocks
 import com.github.hotm.blockentity.AbstractDependableAuraNodeBlockEntity
 import com.github.hotm.blockentity.HotMBlockEntities
 import com.github.hotm.blockentity.PortalReceiverAuraNodeBlockEntity
-import com.github.hotm.world.auranet.AuraNode
-import com.github.hotm.world.auranet.AuraNodeType
-import com.github.hotm.world.auranet.PortalAuraNodeUtils
-import com.github.hotm.world.auranet.PortalReceiverAuraNode
+import com.github.hotm.auranet.AuraNode
+import com.github.hotm.auranet.AuraNodeType
+import com.github.hotm.auranet.PortalAuraNodeUtils
+import com.github.hotm.auranet.PortalReceiverAuraNode
 import com.github.hotm.world.auranet.server.ServerAuraNetStorage
 import net.minecraft.block.BlockRenderType
 import net.minecraft.block.BlockState
