+++ src/main/kotlin/com/github/hotm/blocks/PortalTransmitterAuraNodeBlock.kt
@@ -2,10 +2,10 @@ package com.github.hotm.blocks
 
 import com.github.hotm.blockentity.HotMBlockEntities
 import com.github.hotm.blockentity.PortalTransmitterAuraNodeBlockEntity
-import com.github.hotm.world.auranet.AuraNode
-import com.github.hotm.world.auranet.AuraNodeType
-import com.github.hotm.world.auranet.PortalAuraNodeUtils
-import com.github.hotm.world.auranet.PortalTransmitterAuraNode
+import com.github.hotm.auranet.AuraNode
+import com.github.hotm.auranet.AuraNodeType
+import com.github.hotm.auranet.PortalAuraNodeUtils
+import com.github.hotm.auranet.PortalTransmitterAuraNode
 import com.github.hotm.world.auranet.server.ServerAuraNetStorage
 import net.minecraft.block.BlockRenderType
 import net.minecraft.block.BlockState
